Created a step by step installation guide for running SIP on Windows
Multiple research groups have approached members of the UW Exercise Psychology Lab with requests for information about Sojourns processing for ActiGraph activity data. The current "README.md" and "INSTALL.md" files did not lead to successful installations of SIP or Sojourns for Mac OS or Windows. This installation guide allows people who are new to programming to setup SIP on a Windows 7 or Windows 10 PC to process ActiGraph activity data in the form of 1secDataTable.csv files.